install-multi-user: prompt if proxies are detected during installation
As noted in #1669, setting proxies statically in nix-daemon launch
configuration is potentially problematic (https://github.com/NixOS/nix/issues/1669#issuecomment-365664682)
as these proxies may be tied to a particular network or may already be
managed and propagated to running services as necessary by the user
(see: https://stackoverflow.com/a/26477515).

This commit changes the behavior of the multi-user Linux installer to,
when creating a systemd service for the nix-daemon, ask the user whether
to propagate proxy env vars that are present at install time to the
daemon's service (rather than unconditionally doing so).

This commit also moves some of the proxy env var logic up into
`install-multi-user.sh` in anticipation of the next commit which
has `install-darwin-multi-user.sh` handle proxy env vars similarly
(adding to the launchctl configuration, with the user's consent).